added validations, only the user can delete themself social network

+++ controllers/network.controller.js
@@ -12,43 +12,47 @@ const getSocialNetworks = async (req, res) => {
 
 const deleteSocialNetwork = async (req, res) => {
     const { id: socialNetworkId } = req.params;
-    try{
+    
+    try {
         const socalNetwork = await SocialNetwork.findById(socialNetworkId);
-        if(!socalNetwork) throw new Error('La red social no existe');
-        await SocialNetwork.findByIdAndDelete( socialNetworkId );
+        if (!socalNetwork) throw new Error('La red social no existe');
+        if (socalNetwork.user.toString() !== req.uid)  throw new Error('No tienes permisos para eliminar esta red social');
+        await SocialNetwork.findByIdAndDelete(socialNetworkId);
         res.status(200).json({
             message: 'Red social eliminada correctamente',
         })
-    }catch(error){
+    } catch (error) {
         console.log(error.message);
-    }
+    } 
 }
 
 const createSocialNetwork = async (req, res) => {
-  let user = await User.findOne({ _id: userId }).populate('networks');
-    if(user.networks.find(network => network.name === name)){
+    const {uid:userId }= req;
+    const {name, url} = req.body;
+    console.log(req.body);
+    let user = await User.findOne({ _id: userId }).populate('networks');
+    if (user.networks.find(network => network.name === name)) {
         return res.status(400).json({
             message: `La red social ${name} ya ha sido registrada`
         });
     }
-         try {
-            user = await User.findOne({ _id: userId });
-            let socialNetwork = new SocialNetwork({
-                name,
-                url,
-                user: user._id
-            });
-            const savedSocialNetwork = await socialNetwork.save();
-            console.log(socialNetwork);
-            user.networks = user.networks.concat(savedSocialNetwork._id);
-            await user.save();
-            return res.status(200).json({
-                message: 'Red social agreada correctamente',
-                ok: true
-            });
-        } catch (error) {
-            console.log(error);
-        }  
+    try {
+        user = await User.findOne({ _id: userId });
+        let socialNetwork = new SocialNetwork({
+            name,
+            url,
+            user: user._id
+        });
+        const savedSocialNetwork = await socialNetwork.save();
+        user.networks = user.networks.concat(savedSocialNetwork._id);
+        await user.save();
+        return res.status(200).json({
+            message: 'Red social agreada correctamente',
+            ok: true
+        });
+    } catch (error) {
+        console.log(error);
+    }
 
 }
